// src/shared/apis/telegro/$schemas/cartListDTO.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Telegro API
 * Telegro API 문서입니다.
 * OpenAPI spec version: 1.0.0
 */
import type { CartResponseDTO } from './cartResponseDTO';

/**
 * 응답 데이터
 */
export interface CartListDTO {
  isLast?: boolean;
  totalPage?: number;
  totalElement?: number;
  totalPrice?: number;
  carts?: CartResponseDTO[];
}