// src/shared/apis/telegro/$schemas/noticeListDTO.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Telegro API
 * Telegro API 문서입니다.
 * OpenAPI spec version: 1.0.0
 */
import type { NoticeDTO } from './noticeDTO';

/**
 * 응답 데이터
 */
export interface NoticeListDTO {
  isLast?: boolean;
  totalPage?: number;
  totalElement?: number;
  notices?: NoticeDTO[];
}