// src/shared/apis/telegro/$schemas/orderDetailDTO.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Telegro API
 * Telegro API 문서입니다.
 * OpenAPI spec version: 1.0.0
 */
import type { CartProductDTO } from './cartProductDTO';
import type { OrderDetailDTOOrderStatus } from './orderDetailDTOOrderStatus';
import type { OrderDetailDTOPaymentMethod } from './orderDetailDTOPaymentMethod';
import type { OrderDetailDTOPaymentStatus } from './orderDetailDTOPaymentStatus';
import type { DeliveryAddressDetailDTO } from './deliveryAddressDetailDTO';
import type { UserOrderInfoDTO } from './userOrderInfoDTO';

export interface OrderDetailDTO {
  /** 주문 id */
  orderId?: number;
  /** 주문 번호 */
  orderNumber?: string;
  /** 상품 List */
  products?: CartProductDTO[];
  /** 주문 날짜 */
  createdAt?: string;
  /** 주문 상태를 나타내는 열거형 */
  orderStatus?: OrderDetailDTOOrderStatus;
  /** 요청사항 */
  request?: string;
  /** 주문 총 가격 */
  amount?: number;
  /** 배송비 */
  shoppingCost?: number;
  /** 결제 방식 */
  paymentMethod?: OrderDetailDTOPaymentMethod;
  /** 결제 상태 */
  paymentStatus?: OrderDetailDTOPaymentStatus;
  deliveryAddress?: DeliveryAddressDetailDTO;
  userInfo?: UserOrderInfoDTO;
}